api_service: Add write-only API stub code.
